Add radios for grouping time series by week/month (#18)
This adds radios that make it easy to switch the visualization between grouping by week or month.

It also changes the CSV/JSON of the time series to include a "day" column instead of a "week" column; the actual grouping is done in vega lite on the page.

It also makes a few other stylistic changes, such as increasing padding/spacing.

We also only show the time series data from the beginning of 2020 in the visualization now, fixing #10.  The CSV/JSON files still include data prior to 2020, though.