upgrade to pacel V2 to fix failing builds (#76)
It was pointed out in #75 that our builds have been failing recently. After testing some things out I think it was that the parcel-bundler package had been deprecated, and there is a new V2 that we needed to upgrade to. While I was at it I made small smaller updates to vega packages, and fixed some other react/typescript type issues in the table that were coming up as well.